armv8: adapt variables to register width in exception handlers

+++ src/arch/armv8/aborts.c
@@ -21,7 +21,7 @@
 #include <arch/psci.h>
 #include <hypercall.h>
 
-typedef void (*abort_handler_t)(uint32_t, uint64_t, uint64_t);
+typedef void (*abort_handler_t)(uint64_t, uint64_t, uint64_t);
 
 void internal_abort_handler(uint64_t gprs[]) {
 
@@ -35,13 +35,13 @@ void internal_abort_handler(uint64_t gprs[]) {
     ERROR("cpu%d internal hypervisor abort - PANIC\n", cpu.id);
 }
 
-void aborts_data_lower(uint32_t iss, uint64_t far, uint64_t il)
+void aborts_data_lower(uint64_t iss, uint64_t far, uint64_t il)
 {
     if (!(iss & ESR_ISS_DA_ISV_BIT) || (iss & ESR_ISS_DA_FnV_BIT)) {
         ERROR("no information to handle data abort (0x%x)", far);
     }
 
-    uint32_t DSFC =
+    uint64_t DSFC =
         bit64_extract(iss, ESR_ISS_DA_DSFC_OFF, ESR_ISS_DA_DSFC_LEN) & (0xf << 2);
 
     if (DSFC != ESR_ISS_DA_DSFC_TRNSLT) {
@@ -77,7 +77,7 @@ void aborts_data_lower(uint32_t iss, uint64_t far, uint64_t il)
     }
 }
 
-void smc64_handler(uint32_t iss, uint64_t far, uint64_t il)
+void smc64_handler(uint64_t iss, uint64_t far, uint64_t il)
 {
     uint64_t smc_fid = cpu.vcpu->regs->x[0];
     uint64_t x1 = cpu.vcpu->regs->x[1];
@@ -97,7 +97,7 @@ void smc64_handler(uint32_t iss, uint64_t far, uint64_t il)
     cpu.vcpu->regs->elr_el2 += pc_step;
 }
 
-void hvc64_handler(uint32_t iss, uint64_t far, uint64_t il)
+void hvc64_handler(uint64_t iss, uint64_t far, uint64_t il)
 {
     uint64_t hvc_fid = cpu.vcpu->regs->x[0];
     uint64_t x1 = cpu.vcpu->regs->x[1];
@@ -114,7 +114,7 @@ void hvc64_handler(uint32_t iss, uint64_t far, uint64_t il)
     vcpu_writereg(cpu.vcpu, 0, ret);
 }
 
-void sysreg_handler(uint32_t iss, uint64_t far, uint64_t il)
+void sysreg_handler(uint64_t iss, uint64_t far, uint64_t il)
 {
     vaddr_t reg_addr = iss & ESR_ISS_SYSREG_ADDR;
     emul_handler_t handler = vm_emul_get_reg(cpu.vcpu->vm, reg_addr);
@@ -146,16 +146,16 @@ abort_handler_t abort_handlers[64] = {[ESR_EC_DALEL] = aborts_data_lower,
 
 void aborts_sync_handler()
 {
-    uint32_t esr = MRS(ESR_EL2);
+    uint64_t esr = MRS(ESR_EL2);
     uint64_t far = MRS(FAR_EL2);
     uint64_t hpfar = MRS(HPFAR_EL2);
     uint64_t ipa_fault_addr = 0;
 
     ipa_fault_addr = (far & 0xFFF) | (hpfar << 8);
 
-    uint32_t ec = bit64_extract(esr, ESR_EC_OFF, ESR_EC_LEN);
-    uint32_t il = bit64_extract(esr, ESR_IL_OFF, ESR_IL_LEN);
-    uint32_t iss = bit64_extract(esr, ESR_ISS_OFF, ESR_ISS_LEN);
+    uint64_t ec = bit64_extract(esr, ESR_EC_OFF, ESR_EC_LEN);
+    uint64_t il = bit64_extract(esr, ESR_IL_OFF, ESR_IL_LEN);
+    uint64_t iss = bit64_extract(esr, ESR_ISS_OFF, ESR_ISS_LEN);
 
     abort_handler_t handler = abort_handlers[ec];
     if (handler)
